# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Perpustakaan PPIC API",
        "description": "REST API untuk Sistem Informasi Perpustakaan Politeknik Penerbangan Indonesia Curug. API ini menyediakan akses ke katalog buku, kategori, dan layanan chatbot AI referensi perpustakaan.",
        "contact": {
            "name": "Perpustakaan PPIC",
            "email": "ppicurug.library@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        },
        {
            "url": "/api",
            "description": "API Server (base)"
        }
    ],
    "paths": {
...
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Detail buku",
                "description": "Mengembalikan detail lengkap satu buku berdasarkan ID, termasuk kategori dan riwayat peminjaman.",
                "operationId": "3184b7ae54e17e29fd5fa3027972790e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID buku",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail buku berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Buku tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update data buku",
                "description": "Memperbarui informasi buku yang sudah ada. **Memerlukan autentikasi dan hak akses admin.**",
                "operationId": "63f8f46953b2690763bfbaa581da0668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "stock"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Judul Buku Diperbarui"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Nama Penulis"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "borrowed",
                                            "maintenance",
                                            "lost"
                                        ]
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Buku berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Hapus buku",
                "description": "Menghapus buku dari koleksi. Tidak dapat dihapus jika masih ada peminjaman aktif. **Memerlukan autentikasi dan hak akses admin.**",
                "operationId": "c286320eba3b96b72b7baf6cd634ac1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buku berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Buku tidak dapat dihapus karena masih dipinjam",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot delete book with active borrowings"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Pengantar Teknik Penerbangan"
                    },
                    "author": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-602-123-456-7",
                        "nullable": true
                    },
                    "publisher": {
                        "type": "string",
                        "example": "Gramedia",
                        "nullable": true
                    },
                    "year": {
                        "type": "integer",
                        "example": 2020,
                        "nullable": true
                    },
                    "pages": {
                        "type": "integer",
                        "example": 350,
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "example": "Indonesia",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Buku pengantar ilmu teknik penerbangan...",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Rak A-12",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 3
                    },
                    "status": {
                        "type": "string",
                        "example": "available",
                        "enum": [
                            "available",
                            "borrowed",
                            "maintenance",
                            "lost"
                        ]
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "subjects": {
                        "type": "string",
                        "example": "Penerbangan, Avionik",
                        "nullable": true
                    },
                    "call_number": {
                        "type": "string",
                        "example": "629.13 BUD p",
                        "nullable": true
                    },
                    "gmd_name": {
                        "type": "string",
                        "example": "Text",
                        "nullable": true
                    },
                    "cover_image": {
                        "type": "string",
                        "example": "covers/book1.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Pesan error di sini"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer untuk akses endpoint yang memerlukan autentikasi",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}